rename twitterAgentTool.ts to twitterAgent.ts
